--- FlowCrypt/Controllers/Msg/Helpers/MsgViewController+Setup.swift
@@ -1,47 +1,0 @@
-//
-//  MsgViewController+Setup.swift
-//  FlowCrypt
-//
-//  Created by Anton Kharchevskyi on 06.11.2019.
-//  Copyright © 2019 FlowCrypt Limited. All rights reserved.
-//
-
-import Foundation
-
-extension MsgViewController {
-    enum Parts: Int, CaseIterable {
-        case sender, subject, text
-
-        var indexPath: IndexPath {
-            IndexPath(row: rawValue, section: 0)
-        }
-    }
-    
-    struct Input {
-        var objMessage = MCOIMAPMessage()
-        var bodyMessage: Data?
-        var path = ""
-    }
-
-    enum MessageAction {
-        case moveToTrash, archive, markAsRead, permanentlyDelete
-
-        var text: String? {
-            switch self {
-            case .moveToTrash: return "email_removed".localized
-            case .archive: return "email_archived".localized
-            case .permanentlyDelete: return "email_deleted".localized
-            case .markAsRead: return nil
-            }
-        }
-
-        var error: String? {
-            switch self {
-            case .moveToTrash: return "error_move_trash".localized
-            case .archive: return "error_archive".localized
-            case .permanentlyDelete: return "error_permanently_delete".localized
-            case .markAsRead: return nil
-            }
-        }
-    }
-}
