+++ FlowCrypt/Controllers/Settings/KeySettings/Key List/Node/KeySettingCellNode.swift
@@ -9,6 +9,7 @@
 import AsyncDisplayKit
 import FlowCryptUI
 
+// TODO: ANTON - Move to FlowCryptUI
 final class KeySettingCellNode: CellNode {
     struct Input {
         let title, subtitle, date: NSAttributedString
