+++ FlowCrypt/Models/Folder Models/FolderViewModel.swift
