+++ FlowCrypt/Models/SignIn Models/AppLinks.swift
