+++ FlowCrypt/Models/SignIn Models/SignInType.swift
